requirements.txt
psutil
pyqrcode
pypng
pyzbar
emoji
youtube_dl
pyyaml
redis
coloredlogs
requests
pytz
cowpy
googletrans
beautifulsoup4
gtts
gtts-token
wordcloud
telethon
pillow
python-magic
pygments
distutils2-py3
speedtest-cli
gitpython
werkzeug
flask
flask_sqlalchemy
flask_login
flask_bcrypt
flask_wtf
wtforms
cheroot